build firehose F2 prerequisites
