+++ .github/workflows/publish-crates.yml
@@ -129,13 +129,16 @@ jobs:
           }
 
           # Order matters: each crate must precede anything that depends on it.
-          # omnigraph-compiler and omnigraph-policy have no internal deps;
-          # omnigraph-engine depends on both; omnigraph-api-types and
-          # omnigraph-cluster depend on engine (+ compiler); server depends on
-          # engine + api-types + cluster + the two leaf crates; cli depends on
-          # everything.
+          # omnigraph-compiler, omnigraph-policy, and omnigraph-storage have no
+          # internal deps; omnigraph-control-authority depends on storage;
+          # omnigraph-engine depends on all four; omnigraph-api-types and
+          # omnigraph-cluster depend on engine (+ compiler, and cluster also on
+          # the two control leaves); server depends on engine + api-types +
+          # cluster; cli depends on everything above it.
           publish_if_new omnigraph-compiler
           publish_if_new omnigraph-policy
+          publish_if_new omnigraph-storage
+          publish_if_new omnigraph-control-authority
           publish_if_new omnigraph-engine
           publish_if_new omnigraph-api-types
           publish_if_new omnigraph-cluster
