+++ crates/omnigraph-cluster/Cargo.toml
@@ -16,6 +16,8 @@ failpoints = ["dep:fail", "fail/failpoints", "omnigraph/failpoints"]
 [dependencies]
 omnigraph-compiler = { path = "../omnigraph-compiler", version = "0.9.0" }
 omnigraph = { package = "omnigraph-engine", path = "../omnigraph", version = "0.9.0" }
+omnigraph-storage = { path = "../omnigraph-storage", version = "0.9.0" }
+omnigraph-control-authority = { path = "../omnigraph-control-authority", version = "0.9.0" }
 fail = { workspace = true, optional = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
