+++ crates/omnigraph-control-authority/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "omnigraph-control-authority"
+version = "0.9.0"
+edition = "2024"
+description = "Cycle-free persisted cluster-lock ownership for Omnigraph."
+license = "MIT"
+repository = "https://github.com/ModernRelay/omnigraph"
+homepage = "https://github.com/ModernRelay/omnigraph"
+documentation = "https://docs.rs/omnigraph-control-authority"
+
+[dependencies]
+omnigraph-storage = { path = "../omnigraph-storage", version = "0.9.0" }
+serde = { workspace = true }
+serde_json = { workspace = true }
+thiserror = { workspace = true }
+time = { workspace = true }
+tokio = { workspace = true }
+ulid = { workspace = true }
+
+[dev-dependencies]
+tempfile = { workspace = true }
