+++ crates/omnigraph-storage/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "omnigraph-storage"
+version = "0.9.0"
+edition = "2024"
+description = "Shared local and object-store control-object storage boundary for Omnigraph."
+license = "MIT"
+repository = "https://github.com/ModernRelay/omnigraph"
+homepage = "https://github.com/ModernRelay/omnigraph"
+documentation = "https://docs.rs/omnigraph-storage"
+
+[dependencies]
+async-trait = { workspace = true }
+futures = { workspace = true }
+object_store = { workspace = true }
+sha2 = { workspace = true }
+thiserror = { workspace = true }
+tokio = { workspace = true }
+url = { workspace = true }
+
+[dev-dependencies]
+tempfile = { workspace = true }
